Split filter order and default pipelines
