+++ matchms/filtering/SpectrumProcessor.py
@@ -7,8 +7,8 @@ import numpy as np
 import pandas as pd
 from tqdm import tqdm
 from matchms import Spectrum
-from matchms.filtering.filter_order_and_default_pipelines import (
-    ALL_FILTERS, FILTER_FUNCTION_NAMES, PREDEFINED_PIPELINES)
+from matchms.filtering.default_pipelines import PREDEFINED_PIPELINES
+from matchms.filtering.filter_order import ALL_FILTERS, FILTER_FUNCTION_NAMES
 from matchms.yaml_file_functions import ordered_dump
