+++ tests/filtering/test_spectrum_processor.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 from matchms import SpectrumProcessor
-from matchms.filtering.filter_order_and_default_pipelines import BASIC_FILTERS
+from matchms.filtering.default_pipelines import BASIC_FILTERS
 from matchms.filtering.SpectrumProcessor import ProcessingReport
 from ..builder_Spectrum import SpectrumBuilder
